+++ src/main/java/com/zhiliao/hotel/utils/TokenUtil.java
@@ -19,12 +19,12 @@ public class TokenUtil {
                 .sign(Algorithm.HMAC256(wxuser.getWxopenid())); // 使用HMAC256生成token
     }
 
-    public Integer getUserId(String token) {
+    public Long getUserId(String token) {
         String userId;
         try {
             // 获取token中的userId
             userId = JWT.decode(token).getAudience().get(0);
-            return Integer.parseInt(userId);
+            return Long.parseLong(userId);
         } catch (JWTDecodeException j) {
             throw new RuntimeException("401 错误token，没有访问权限，请重新登录");
         }
